fix(NFTFeature): uploading pull request changes
