+++ src/assets/scss/elements/_product-details.scss
@@ -29,7 +29,6 @@
     border-radius: 20px;
     height: auto;
     flex-basis: 75%;
-    /* padding-left: 20px; */
 
     .tab-pane {
       height: 100%;
