+++ src/components/item-details/bid-tab/details-tab-content.tsx
