+++ src/components/item-details/category.tsx
