+++ src/components/item-details/gallery-tab.tsx
