+++ src/components/top-seller/top-seller-2.tsx
@@ -30,7 +30,7 @@ const TopSeller2 = ({ name, time, path, image, eth, isVerified }: Props) => (
       )}
       <div className="top-seller-content">
         <span>
-          {eth && <>{`${eth} JUP`} by</>}
+          {eth && <>{`${eth} ETH`} by</>}
           <Anchor path={path}>{name}</Anchor>
         </span>
         {time && <span className="count-number">{time}</span>}
