+++ src/pages/_document.tsx
@@ -10,7 +10,7 @@ class MyDocument extends Document {
     return (
       <Html>
         <Head />
-        <body className="">
+        <body>
           <Main />
           <NextScript />
         </body>
