+++ src/pages/index.tsx
@@ -5,7 +5,7 @@ import Header from '@layout/header';
 import Footer from '@layout/footer';
 import HeroArea from '@containers/hero';
 import ServiceArea from '@containers/services';
-import NewestItmesArea from '@containers/product/newest-item';
+import NewestItmesArea from '@containers/item/newest-item';
 import { useEffect } from 'react';
 import homepageData from '../data/homepages/home-01.json';
 import useAppDispatch from '../store/hooks/useAppDispatch';
